requirements.txt
chromadb>=0.4.0
llama-index-core>=0.10.0
llama-index-llms-google-genai>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
google-generativeai>=0.3.0
google-api-core>=2.15.0
streamlit>=1.28.0
nest-asyncio>=1.5.0